+++ pynenc/trigger/__init__.py
@@ -1,5 +1,4 @@
 from pynenc.trigger.base_trigger import BaseTrigger
-from pynenc.trigger.disabled_trigger import DisabledTrigger
 from pynenc.trigger.mem_trigger import MemTrigger
 from pynenc.trigger.sqlite_trigger import SQLiteTrigger
 from pynenc.trigger.trigger_builder import TriggerBuilder
@@ -8,7 +7,6 @@ from pynenc.trigger.types import ConditionId, TriggerDefinitionId
 
 __all__ = [
     "BaseTrigger",
-    "DisabledTrigger",
     "MemTrigger",
     "SQLiteTrigger",
     "TriggerBuilder",